requirements.txt
boto3==1.18.17
botocore==1.21.17
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
idna==3.2
imgkit==1.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
praw==7.4.0
prawcore==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
update-checker==0.18.0
urllib3==1.26.6
websocket-client==1.1.1
Werkzeug==2.0.1
